Split up hosttech_dns module_utils.
